Remove morphware code
