+++ tests/queries/test_abi_query.py
@@ -1,39 +1,8 @@
 from telliot_feeds.queries.abi_query import AbiQuery
-from telliot_feeds.queries.morphware import Morphware
-
-
-def test_query_data():
-    q = Morphware(version=1)
-    print(q.query_data.hex())
-    print(q.abi)
-    print(q.query_id.hex())
-
-    qr = AbiQuery.get_query_from_data(q.query_data)
-    assert isinstance(qr, Morphware)
-    assert qr.version == 1
 
 
 def test_get_query_from_data():
-    query_data = (
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00@\x00"
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x80\x00"
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\tMorphwar"
-        b"e\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-        b"\x00\x00\x00\x00\x00\x00 \x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-        b"\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00"
-        b"\x00\x00\x00\x00\x00\x01"
-    )
-
-    q = AbiQuery.get_query_from_data(query_data)
-    print(q)
-
-    assert isinstance(q, Morphware)
-    assert q.version == 1
+    pass
 
 
 def test_generate_query_data():
