+++ tests/utils/test_reporter_utils.py
@@ -54,7 +54,6 @@ def test_reporter_sync_schedule_list():
     lis = reporter_sync_schedule
     assert len(lis) > 4
     assert "eth-usd-spot" in lis
-    assert "morphware-v1" not in lis
     assert "uspce-legacy" not in lis
